feat: test transactions on the dev chain

+++ tx-builder/src/ckb/metadata.rs
@@ -66,7 +66,7 @@ where
             self.last_checkpoint
                 .propose_count
                 .iter()
-                .map(|v| (v.proposer.0.into(), v.count as u64))
+                .map(|v| (v.proposer.0.into(), v.count))
                 .collect(),
         )
         .await?;
